moved tests from AdvancedFormatting to lifecycleFormEditor
